requirements.txt
Django==4.2.*
environs==9.5.*
Pillow==10.1.*
django-admin-sortable2~=2.1.10
django-tinymce~=3.6.1